requirements.txt
asgiref             3.10.0
azure-core          1.36.0
azure-storage-blob  12.27.1
certifi             2025.10.5
cffi                2.0.0
charset-normalizer  3.4.4
cryptography        46.0.3
Django              5.2.7
django-axes         8.0.0
django-recaptcha    4.1.0
django-sendgrid-v5  1.3.0
django-storages     1.14.6
djangorestframework 3.16.1
gunicorn            23.0.0
idna                3.11
isodate             0.7.2
MarkupSafe          3.0.3
mysqlclient         2.2.7
narwhals            2.10.2
numpy               2.2.6
packaging           25.0
pandas              2.3.3
pillow              12.0.0
pip                 22.0.2
plotly              6.4.0
pycparser           2.23
python-dateutil     2.9.0.post0
python-decouple     3.8
python-http-client  3.3.7
pytz                2025.2
requests            2.32.5
sendgrid            6.12.5
setuptools          59.6.0
six                 1.17.0
sqlparse            0.5.3
typing_extensions   4.15.0
tzdata              2025.2
urllib3             2.5.0
Werkzeug            3.1.3
whitenoise          6.11.0